+++ drv2605.h
@@ -70,6 +70,6 @@ void set_realtimeValue(uint8_t rtp);
 void use_ERM();
 void use_LRA();
 
-const char* identify_effect(int effect);
+const char* identify_effect(uint8_t effect);
 
 #endif
